NeuroMotion_Pi/pyproject.toml
[project]
name = "DARKSide" 
version = "0.1.0"
description = "Parkinson's Disease Detection"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "numpy>=1.21.0",
    "scipy>=1.7.0",
    "paho-mqtt>=1.6.1",
    "pyyaml>=6.0",
    "smbus2>=0.5.0",
    "adafruit-circuitpython-adxl34x>=1.12.15",
]

[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "ruff>=0.11.3",
]

[project.optional-dependencies]
dev = ["black", "isort"]
